Old_Code/requirements-flexible.txt
# More flexible versions
streamlit>=1.25.0
opencv-python>=4.5.0
Pillow>=9.0.0
pytesseract>=0.3.8
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
torch>=1.12.0
transformers>=4.20.0
sentence-transformers>=2.0.0
scikit-learn>=1.1.0